test(meshtastic): consolidate utility and coverage tests
Merge fragmented test files into primary utility test suites to reduce file sprawl and simplify the test directory structure.

Changes:
- Absorbed async helper tests and dummy loop implementations from `test_meshtastic_utils_async_helpers.py` into `test_meshtastic_utils_async.py`.
- Integrated event guard coverage tests from `test_meshtastic_utils_event_guards_coverage.py` into `test_meshtastic_utils_message_edge.py`.
- Updated `reset_meshtastic_relay_state` fixtures to include additional global state resets for client and connection flags.
- Normalized imports and added missing constants to support the consolidated test cases.
